requirements.txt
flask==2.2.2
monzo-api==0.2.0
